# specs/003-schema-redesign/contracts/api.yaml
openapi: 3.0.3
info:
  title: Jump Ticketing API — Schema Redesign
  description: |
    REST API contracts for the schema redesign feature (003).
    Covers new endpoints for Organizations, Venues, Price Tiers, Orders,
    Users, and Guest Lookup. Updates existing Event and Ticket endpoints.
    Authentication moves to Auth.js v5 (JWT in Authorization header).
  version: 2.0.0
  contact:
    name: Jump Platform
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.jump.example.com/v1
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /organizations:
    get:
      tags: [Organizations]
      summary: List organizations
      operationId: listOrganizations
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Organizations retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Organization"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Organizations]
      summary: Create an organization
      operationId: createOrganization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationInput"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}:
    get:
      tags: [Organizations]
      summary: Get organization details
      operationId: getOrganization
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Organizations]
      summary: Update an organization
      operationId: updateOrganization
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationInput"
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===== VENUES (FR-049) =====
...
  /organizations/{organizationId}/events/{eventId}/price-tiers:
    get:
      tags: [PriceTiers]
      summary: List price tiers for an event
      operationId: listPriceTiers
      security: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/EventIdParam"
      responses:
        "200":
          description: Price tiers
          content:
            application/json:
              schema:
                type: object
                properties:
                  priceTiers:
                    type: array
                    items:
                      $ref: "#/components/schemas/PriceTier"
    post:
      tags: [PriceTiers]
      summary: Add a price tier to an event
      operationId: createPriceTier
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/EventIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePriceTierInput"
      responses:
        "201":
          description: Price tier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceTier"
        "400":
          description: Validation error (e.g., total tier inventory exceeds event capacity)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organizations/{organizationId}/events/{eventId}/price-tiers/{priceTierId}:
    patch:
      tags: [PriceTiers]
      summary: Update a price tier
      operationId: updatePriceTier
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/EventIdParam"
        - $ref: "#/components/parameters/PriceTierIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePriceTierInput"
      responses:
        "200":
          description: Price tier updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceTier"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/events/{eventId}/price-tiers/{priceTierId}/activate:
    post:
      tags: [PriceTiers]
      summary: Activate a price tier
      operationId: activatePriceTier
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/EventIdParam"
        - $ref: "#/components/parameters/PriceTierIdParam"
      responses:
        "200":
          description: Price tier activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceTier"
  /organizations/{organizationId}/events/{eventId}/price-tiers/{priceTierId}/deactivate:
    post:
      tags: [PriceTiers]
      summary: Deactivate a price tier
      operationId: deactivatePriceTier
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/EventIdParam"
        - $ref: "#/components/parameters/PriceTierIdParam"
      responses:
        "200":
          description: Price tier deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceTier"
  /organizations/{organizationId}/events/{eventId}/price-tiers/reorder:
    post:
      tags: [PriceTiers]
      summary: Reorder price tiers
      operationId: reorderPriceTiers
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/EventIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tierIds]
              properties:
                tierIds:
                  type: array
                  items:
                    type: string
                  description: Ordered list of price tier IDs
      responses:
        "200":
          description: Tiers reordered
          content:
            application/json:
              schema:
                type: object
                properties:
                  priceTiers:
                    type: array
                    items:
                      $ref: "#/components/schemas/PriceTier"

  # ===== ORDERS (FR-052, FR-053, FR-054) =====
...
  /tickets/redeem:
    post:
      tags: [Tickets]
      summary: Redeem a ticket via QR code scan
      description: |
        Accepts the QR code JWT payload, validates signature, checks
        event match, verifies ticket status. On success, marks VALID → REDEEMED
        with timestamp. Lazy expiration: if event has passed, returns EXPIRED
        and updates status (FR-032).
      operationId: redeemTicket
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemTicketInput"
      responses:
        "200":
          description: Ticket redeemed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedemptionResult"
        "400":
          description: Invalid QR code (forged, tampered, malformed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Ticket already redeemed, expired, voided, or wrong event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedemptionRejection"

  # ===== USERS (FR-056) =====
  /users:
    get:
      tags: [Users]
      summary: List users (admin-only)
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: role
          in: query
          schema:
            type: string
            enum: [CUSTOMER, ORGANIZER, ADMIN]
        - name: organizationId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserSummary"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /users/{userId}:
    patch:
      tags: [Users]
      summary: Update user role or status (admin-only)
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSummary"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===== ANALYTICS (FR-057) =====
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth.js v5 JWT (HS256). Payload includes sub (userId), email, role.
  parameters:
    OrganizationIdParam:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
...
    EventIdParam:
      name: eventId
      in: path
      required: true
      schema:
        type: string
    PriceTierIdParam:
      name: priceTierId
      in: path
      required: true
      schema:
        type: string
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient role or organization access
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: "Tier inventory exceeds event capacity"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # ─── Organization ───
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateOrganizationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
    UpdateOrganizationInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        status:
          type: string
          enum: [ACTIVE, INACTIVE]

    # ─── Venue ───
...
    PriceTier:
      type: object
      properties:
        id:
          type: string
        eventId:
          type: string
        name:
          type: string
        price:
          type: number
          format: decimal
        quantityTotal:
          type: integer
        quantitySold:
          type: integer
        quantityReserved:
          type: integer
        quantityAvailable:
          type: integer
          description: "Computed: quantityTotal - quantitySold - quantityReserved"
        displayOrder:
          type: integer
        minPerOrder:
          type: integer
          nullable: true
        maxPerOrder:
          type: integer
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePriceTierInput:
      type: object
      required: [name, price, quantityTotal]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        price:
          type: number
          minimum: 0
        quantityTotal:
          type: integer
          minimum: 1
        displayOrder:
          type: integer
        minPerOrder:
          type: integer
          minimum: 1
        maxPerOrder:
          type: integer
          minimum: 1
    UpdatePriceTierInput:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
          minimum: 0
        quantityTotal:
          type: integer
          minimum: 1
        displayOrder:
          type: integer
        minPerOrder:
          type: integer
          nullable: true
        maxPerOrder:
          type: integer
          nullable: true

    # ─── Order ───
...
    RedeemTicketInput:
      type: object
      required: [qrPayload]
      properties:
        qrPayload:
          type: string
          description: Raw JWT string scanned from QR code
        eventId:
          type: string
          description: Expected event ID (for cross-event validation)
    RedemptionResult:
      type: object
      properties:
        status:
          type: string
          enum: [REDEEMED]
        ticketId:
          type: string
        barcode:
          type: string
        priceTierName:
          type: string
        contactName:
          type: string
        redeemedAt:
          type: string
          format: date-time
    RedemptionRejection:
      type: object
      properties:
        status:
          type: string
          enum: [ALREADY_REDEEMED, EXPIRED, VOIDED, WRONG_EVENT]
        ticketId:
          type: string
        message:
          type: string
        originalRedemptionTime:
          type: string
          format: date-time
          nullable: true
          description: Only present for ALREADY_REDEEMED

    # ─── Payment ───
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        role:
          type: string
          enum: [CUSTOMER, ORGANIZER, ADMIN]
        organizationId:
          type: string
          nullable: true
        organizationName:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UpdateUserInput:
      type: object
      properties:
        role:
          type: string
          enum: [CUSTOMER, ORGANIZER, ADMIN]
        organizationId:
          type: string
          nullable: true
        isActive:
          type: boolean

    # ─── Analytics ───
...